+++ src/Plugin.php
@@ -81,14 +81,10 @@ class Plugin extends BasePlugin
         parent::init();
         self::$plugin = $this;
         $this->_registerHasCpSection();
+        $this->_registerProjectConfigRebuild();
 
         if ($this->getSettings()->useProjectConfig) {
             $this->_registerProjectConfigApply();
-            $this->_registerProjectConfigRebuild();
-        }
-
-        if (Craft::$app->getConfig()->getGeneral()->allowAdminChanges) {
-            $this->_registerUseProjectConfigSettingCheck();
         }
 
         if (Craft::$app->getRequest()->getIsCpRequest()) {
@@ -152,47 +148,10 @@ class Plugin extends BasePlugin
     private function _registerProjectConfigRebuild(): void
     {
         Event::on(ProjectConfig::class, ProjectConfig::EVENT_REBUILD, function(RebuildConfigEvent $event) {
-            $event->config['contentTemplates'] = $this->_configFromDb();
+            $event->config['contentTemplates'] = $this->getSettings()->useProjectConfig
+                ? $this->projectConfig->getFromDb()
+                : null;
         });
-    }
-
-    /**
-     * Adds or removes all Content Templates project config data if the `useProjectConfig` plugin setting has recently been changed.
-     */
-    private function _registerUseProjectConfigSettingCheck(): void
-    {
-        Craft::$app->on(Application::EVENT_INIT, function() {
-            $projectConfig = Craft::$app->getProjectConfig();
-            $useProjectConfig = $this->getSettings()->useProjectConfig;
-            $hasProjectConfig = $projectConfig->get('contentTemplates');
-
-            if ($useProjectConfig && !$hasProjectConfig) {
-                $projectConfig->set('contentTemplates', $this->_configFromDb());
-            } elseif (!$useProjectConfig && $hasProjectConfig) {
-                $projectConfig->remove('contentTemplates');
-            }
-        });
-    }
-
-    private function _configFromDb(): array
-    {
-        $contentTemplateConfig = [];
-        $contentTemplateOrdersConfig = [];
-
-        foreach (ContentTemplate::find()->withStructure(true)->all() as $contentTemplate) {
-            $config = $contentTemplate->getConfig();
-            $contentTemplateConfig[$contentTemplate->uid] = $config;
-            $contentTemplateOrdersConfig[$config['type']][$config['sortOrder']] = $contentTemplate->uid;
-        }
-
-        foreach ($contentTemplateOrdersConfig as $typeUid => $templateUids) {
-            $contentTemplateOrdersConfig[$typeUid] = array_values($templateUids);
-        }
-
-        return [
-            'templates' => $contentTemplateConfig,
-            'orders' => $contentTemplateOrdersConfig,
-        ];
     }
 
     /**