+++ src/console/controllers/PcController.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace spicyweb\contenttemplates\console\controllers;
+
+/**
+ * Actions for managing Content Templates project config data.
+ *
+ * @package spicyweb\contenttemplates\console\controllers
+ * @author Spicy Web <plugins@spicyweb.com.au>
+ * @since 2.1.0
+ */
+class PcController extends ProjectConfigController
+{
+}